Extract Environment
