+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/Project/Environment/elements.ts
@@ -1,0 +1,10 @@
+import styled, { css } from 'styled-components';
+
+export const BundlerLink = styled.a.attrs({
+  target: '_blank',
+  rel: 'noreferrer noopener',
+})`
+  ${({ theme }) => css`
+    color: ${theme.templateColor} !important;
+  `}
+`;
